Contacts Section Bug Fix
Major bug finally fix with having to double click the letter to get the
menu open after the initial press/open of a previous letter. THe problem
was not the ltnBtn loop but the cleanup was finding ALL contacts items
and deleting them ALL. Added newItems <List> To hold only the current
items of letter selected. Problem solved.