Cell Phone Contacts Update
BUG: The other Letter Panels are not moving down when a single leter is
opened with contacts list. The list can not be clicked as clicking opens
another letter. Need to figure out how to update list with new space.
Already created a size and distance calculator that dynamically asses
the list panel size (height) need and or distance it would need to move.
Moving on to another section today. Will return to this issue whn I know
more about it. Research required :/