Day Night Cycle Implementation
Started the day and night cycle implementation for the desk window.
Started the system that will govern the time scale. Also added the
ability for the user to close the blinds and a computer has been added
to the main desk scene. This will have a cd tray that opens for the user
to watch training videos.