Dialogue System Implemented for NPCs
Made a new project to work on dialogue system. It is 95% complete just a
couple bugs. You can now put an NPC in the world as a button and it will
function when clicked. A dialogue window pops with player dialogue
options at the bottom. XML has been added to control the flow of NPC,
Dialogue and Options nodes. Working on changing the way the NPC vs
Dialogue node is selected. The final goal is a perfect object to
automate dialogues in game.